+++ src/pages/CreatePage/CreateComponents/BackGround/BackGround.jsx
@@ -1,14 +1,14 @@
 import { useState } from 'react';
 import styles from './BackGround.module.css';
 
-import img1 from '../../../assets/images/background_green.png';
-import img2 from '../../../assets/images/background_yellow.png';
-import img3 from '../../../assets/images/background_blue.png';
-import img4 from '../../../assets/images/background_pink.png';
-import img5 from '../../../assets/images/img5.png';
-import img6 from '../../../assets/images/img6.png';
-import img7 from '../../../assets/images/img7.png';
-import img8 from '../../../assets/images/img8.png';
+import img1 from '../../../../assets/images/background_green.png';
+import img2 from '../../../../assets/images/background_yellow.png';
+import img3 from '../../../../assets/images/background_blue.png';
+import img4 from '../../../../assets/images/background_pink.png';
+import img5 from '../../../../assets/images/img5.png';
+import img6 from '../../../../assets/images/img6.png';
+import img7 from '../../../../assets/images/img7.png';
+import img8 from '../../../../assets/images/img8.png';
 
 const images = [
   { id: 'green', src: img1 },
