+++ src/pages/CreatePage/CreateComponents/PasswordCheck.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import styles from '../input/Input.module.css';
+import styles from '../../../components/input/Input.module.css';
 
-import closeeye from '../../assets/icons/ic_close_eye.png';
-import openeye from '../../assets/icons/ic_eye.png';
+import closeeye from '../../../assets/icons/ic_close_eye.png';
+import openeye from '../../../assets/icons/ic_eye.png';
 
 const PasswordCheck = ({ password, passwordCheck, setPasswordCheck }) => {
   const [error, setError] = useState('');
