+++ src/pages/CreatePage/CreateComponents/StudyName.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import styles from '../input/Input.module.css';
+import styles from '../../../components/input/Input.module.css';
 
 const StudyName = ({ title, setTitle }) => {
   const [error, setError] = useState('');
